requirements.txt
requests
beautifulsoup4
pyrogram
TgCrypto
python-dotenv